\name{RemoveUnrecognizeAminoAcid}
\alias{RemoveUnrecognizeAminoAcid}
\title{RemoveUnrecognizeAminoAcid}
\description{
This function remove amino acid code that is not in these 20 amino acids: Glycine/G, Alanine/A, Valine/V, Cysteine/C, Proline/P, Leucine/L, Isoleucine/I, Methionine/M, Tryptophan/W, Phenylalanine/F, Lysine/K, Arginine/R, Histidine/H, Serine/S, Threonine/T, Tyrosine/Y, Asparagine/N, Glutamine/Q, Aspartic Acid/D, Glutamic Acid/E
}
\usage{
RemoveUnrecognizeAminoAcid(protein_sequence)
}
\examples{
RemoveUnrecognizeAminoAcid("RSVTTEINTLFQTLTSIAEKVDP")
}
